refactor: move create[..]Ref function into new file
